[Prover Service] Rename jwk_fetching -> jwk_fetcher.
