+++ prover-service/src/external_resources/mod.rs
@@ -1,4 +1,4 @@
 // Copyright (c) Aptos Foundation
 
-pub mod jwk_fetching;
+pub mod jwk_fetcher;
 pub mod prover_config;
