+++ prover-service/src/main.rs
@@ -7,7 +7,7 @@ use clap::Parser;
 use hyper::service::{make_service_fn, service_fn};
 use hyper::Server;
 use prover_service::external_resources::prover_config::ProverServiceConfig;
-use prover_service::external_resources::{jwk_fetching, prover_config};
+use prover_service::external_resources::{jwk_fetcher, prover_config};
 use prover_service::request_handler::prover_state::{ProverServiceState, TrainingWheelsKeyPair};
 use prover_service::request_handler::{deployment_information, handler};
 use prover_service::*;
@@ -59,7 +59,7 @@ async fn main() {
     load_verification_key(prover_service_config.clone());
 
     // init jwk fetching job; refresh every `config.jwk_refresh_rate_secs` seconds
-    jwk_fetching::init_jwk_fetching(
+    jwk_fetcher::init_jwk_fetching(
         &prover_service_config.oidc_providers,
         Duration::from_secs(prover_service_config.jwk_refresh_rate_secs),
     )
